sophisticated contentSchema
