+++ models/dbSchemas.ts
@@ -19,27 +19,78 @@ enum contentColumn {
 };
 
 const userSchema = new Schema({
-    userID: Number,
-    name: String,
-    email: String,
-    profilePicture: String, // String -- maybe imgr
-    role: userRole,
-    submissions: Array<Number>
+    userID: {
+        type: Number,
+        required: true
+    },
+    name: {
+        type: String,
+        required: true
+    },
+    email: {
+        type: String,
+        required: true
+    },
+    profilePicture: {
+        type: String,
+        required: true
+    },
+    role: {
+        type: userRole,
+        default: userRole.Reader
+    },
+    submissions: {
+        type: Array<Schema.Types.ObjectId>,
+        ref: "Content"
+    }
 });
 
 const contentSchema = new Schema({
-    userID: Number,
-    contentID: Number,
-    title: String,
-    author: String,
-    nodeType: contentType,
-    textContent: String,
-    imgContent: String, // imgr
-    vidConent: String, //youtube?
-    column: contentColumn,
-    dateCreated: Date,
-    lastUpdated: Date
+    userID: {
+        type: Number,
+        required: true
+    },
+    contentID: {
+        type: Number,
+        required: true
+    },
+    title: {
+        type: String,
+        required: true
+    },
+    author: {
+        type: Schema.Types.ObjectId,
+        ref: "User"
+    },
+    nodeType: {
+        type: contentType,
+        required: true
+    },
+    textContent: {
+        type: String,
+        required: false
+    },
+    imgContent: {
+        type: String,
+        required: false
+    },
+    vidContent: {
+        type: String,
+        required: false
+    },
+    column: {
+        type: contentColumn,
+        required: true
+    },
+    dateCreated: {
+        type: Date,
+        default: Date.now
+    },
+    lastUpdated: {
+        type: Date,
+        default: Date.now
+    }
 });
 
 export const User = model("User", userSchema);
-export const Content = model("`Content", contentSchema);
+export const Content = model("Content", contentSchema);
